python/py_jpeg_converter/pyproject.toml
[project]
name = "jpeg_converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "ipykernel>=6.29.5",
  "pillow>=11.1.0",
  "pillow-heif>=0.21.0",
  "pytest>=8.3.5",
  "pytest-mock>=3.14.0",
  "rich>=13.9.4",
  "typer>=0.15.2",
]
